resin-image.inc: Update image to match new bsp layers
The new bsp layers have different names for the device
tree files and the image flasher binary.
Firmware-imx-brcm is not present in the new layers
and has as a replaceable candidate firmware-imx.
RESIN_BOOT_PARTITION_FILES was updated with the new
dtbs, IMAGE_INSTALL_APPEND was updates with the new
firmware-imx and IMAGE_CMD with the new flasher binary.

Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>