conf/recipes-kernel/recipes-support: Add frequency scaling governor service
The service will monitor the CPU temperature and when
it is over a threshold value (75 degrees) it will set
the CPU governor to powersave.
The schedutil scaling governor has been added to kernel
config.
If the CPU temperature will drop below the threshold value,
the CPU governor will be set to schedutil

Changelog-entry: Add service to set CPU governor by temperature
Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>