linux-imx: Disable internal imx RTC and external i2c RTC
The current hardware setup for the EP does not contain a coin
battery for the internal RTC on the SoM. The system queries the
device created by the RTC and the retrieved date and time value
is used by chronyd. This leads to chronyd using wrong (default)
date and time value from the RTC. To fix this the internal RTC
will be disabled

On the EP board we don't have an external RTC on i2c, so that
needs to be disabled.

Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>