linux-imx: Add sw version specific DTBs
The EP board had 3 major hardware updates
and those needed to be configured in software (OS)
to have the same OS image for all hardware versions.

The versions where software was updated due to hardware
changes were v2.46.0+rev10 and v2.51.1+rev3.
The DTBs for those software versions were loaded in the
deploy dir of the build and made available for the
balena boot partition

Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>