linux-imx: goodix touch panel - add support for ESD
Add ESD (Electrostatic Discharge) protection mechanism.

The driver enables ESD protection in HW and checks a register
to determine if ESD occurred. If ESD is signalled by the HW,
the driver will reset the device.

The ESD poll time (in ms) can be set through
esd-recovery-timeout-ms ACPI/DT property. If it is set to 0,
ESD protection is disabled.

Changelog-entry: Add touchpanel ESD support
Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>