recipes-kernel: Disable RTCs for all EP hw versions
The RTC was disabled in device tree of the main (latest) hw
version of the EP  because the time and date are taken
from an NTP server.
The DTBs which are precompiled and are for older EP hw
versions were omitted. This commit updates those DTBs

Changelog-entry: Disable RTCs for all EP hw versions
Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>