local.conf.sample: Update host tools to replace docker with balena-engine
Add a set of host tools dependencies to use rootless balena-engine.

For overlay based devices, the docker host tool will be removed once the
meta-balena supports building with balena-engine.

For devcies that use aufs we leave docker as rootless is not supported over
aufs, and we will remove it once aufs devices are deprecated.

Change-type: patch
Changelog-entry: Update host tools to replace docker with balena-engine in build process
Signed-off-by: Alex Gonzalez <alexg@balena.io>